sid-backend/requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-cache2==0.2.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pendulum==2.1.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==1.10.12
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
